+++ pkg/integrations/squid_exporter/squid_exporter_test.go
@@ -1,0 +1,1 @@
+package squid_exporter